SRGAN3/requirements.txt
backports.entry-points-selectable==1.1.0
certifi==2022.9.24
charset-normalizer==2.1.1
distlib==0.3.2
filelock==3.0.12
idna==3.4
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
packaging==21.3
pathlib==1.0.1
Pillow==9.3.0
platformdirs==2.2.0
pyparsing==3.0.9
requests==2.28.1
six==1.16.0
torch==1.13.0
torchmetrics==0.11.0
torchvision==0.14.0
typing-extensions==4.4.0
urllib3==1.26.13
virtualenv==20.7.0
